requirements.txt
astroid==2.3.2
autopep8==1.4.4
boto==2.49.0
boto3==1.10.4
botocore==1.13.4
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-appconf==1.0.3
django-crispy-forms==1.8.0
django-dotenv==1.4.2
django-froala-editor==3.0.6
django-storages==1.7.2
djangorestframework==3.10.3
docutils==0.15.2
gunicorn==19.9.0
idna==2.8
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.4
pycodestyle==2.5.0
pylint==2.4.3
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
requests-toolbelt==0.9.1
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
Wand==0.5.7
whitenoise==4.1.4
wrapt==1.11.2
